+++ lib/const/app_paint.dart
@@ -7,12 +7,16 @@ class AppPaint {
   static const BLACK = Color.fromARGB(255, 0, 0, 0);
   static const GREY = Color.fromARGB(255, 66, 66, 66);
   static const BLUE_DARK = Color.fromARGB(255, 68, 68, 255);
+  static const BLUE_LIGHT = Color.fromARGB(187, 68, 68, 255);
+
   static const TRANSPARENT = Color.fromARGB(0, 0, 0, 0);
   static const GREY_LIGHT = Color.fromARGB(255, 238, 238, 238);
 
   static const RED_DARK = Color.fromARGB(255, 198, 56, 45);
+  static const RED_LOW_LIGHT = Color.fromARGB(137, 255, 134, 126);
+  static const RED_MODERATE_LIGHT = Color.fromARGB(204, 198, 55, 45);
+
   static const YELLOW_LIGHT = Color.fromARGB(255, 255, 245, 157);
-  static const RED_MODERATE_LIGHT = Color.fromARGB(137, 255, 134, 126);
 
   static const PURPLE_LIGHT = Color.fromARGB(120, 160, 126, 255);
   static const AMBER_LIGHT = Color.fromARGB(211, 255, 193, 7);
