+++ lib/todo_view/todo_view_export.dart
@@ -1,7 +1,9 @@
-export 'todo_screen.dart';
 export 'components/cupertino_popup_design.dart';
-export 'components/no_todo_body_view.dart';
-
 export 'components/functions/cupertino_popup_add_todo_functionality.dart';
 export 'components/functions/cupertino_popup_edit_todo_functionality.dart';
 export 'components/functions/cupertino_popup.dart';
+export 'todo_screen.dart';
+export 'pending_todo_view.dart';
+export 'done_todo_view.dart';
+export 'components/linear_decor_text.dart';
+export 'components/app_bar_logo.dart';
